move primitives out of runtime
