+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
 	"pallets/root-of-trust/rpc",
 	"pallets/root-of-trust/rpc/runtime-api",
 	"pallets/tcr",
+	"primitives",
 	"runtime",
 	"support"
 ]
