+++ node/Cargo.toml
@@ -26,6 +26,7 @@ frame-benchmarking-cli = { version = "2.0.0-rc6", default-features = false }
 futures = { version = "0.3.5", features = ["compat"] }
 jsonrpc-core = "14.2.0"
 jsonrpc-pubsub = "14.2.0"
+nodle-chain-primitives = { version = "2.0.0-rc6", path = "../primitives" }
 nodle-chain-runtime = { version = "2.0.0-rc6", path = "../runtime" }
 pallet-im-online = { version = "2.0.0-rc6", default-features = false }
 pallet-root-of-trust-rpc  = { version = "2.0.0-rc6", path = "../pallets/root-of-trust/rpc" }
