+++ node/src/command.rs
@@ -21,7 +21,7 @@ use crate::{
     cli::{Cli, Subcommand},
     service::{self, new_partial, Executor},
 };
-use nodle_chain_runtime::Block;
+use nodle_chain_primitives::Block;
 use sc_cli::{ChainSpec, Result, Role, RuntimeVersion, SubstrateCli};
 use sc_service::PartialComponents;
