+++ primitives/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "nodle-chain-primitives"
+version = "2.0.0-rc6"
+authors = ["Eliott Teissonniere <git.eliott@teissonniere.org>"]
+edition = "2018"
+
+[dependencies]
+frame-system = { version = "2.0.0-rc6", default-features = false }
+parity-scale-codec = { version = "1.2.0", default-features = false, features = ["derive"] }
+sp-application-crypto = { version = "2.0.0-rc6", default-features = false }
+sp-core = { version = "2.0.0-rc6", default-features = false }
+sp-runtime = { version = "2.0.0-rc6", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+	"frame-system/std",
+    "parity-scale-codec/std",
+	"sp-application-crypto/std",
+	"sp-core/std",
+	"sp-runtime/std",
+]
