+++ runtime/Cargo.toml
@@ -10,6 +10,7 @@ std = [
   "frame-executive/std",
   "frame-support/std",
   "frame-system/std",
+  "nodle-chain-primitives/std",
   "frame-system-rpc-runtime-api/std",
   "pallet-allocations/std",
   "pallet-amendments/std",
@@ -96,6 +97,7 @@ frame-support = { version = "2.0.0-rc6", default-features = false }
 frame-system = { version = "2.0.0-rc6", default-features = false }
 frame-system-benchmarking = { version = "2.0.0-rc6", default-features = false, optional = true }
 frame-system-rpc-runtime-api = { version = "2.0.0-rc6", default-features = false }
+nodle-chain-primitives = { version = "2.0.0-rc6", default-features = false, path = "../primitives" }
 pallet-allocations = { version = "2.0.0-rc6", default-features = false, path = "../pallets/allocations" }
 pallet-amendments = { version = "2.0.0-rc6", default-features = false, path = "../pallets/amendments" }
 pallet-authority-discovery = { version = "2.0.0-rc6", default-features = false }
@@ -146,6 +148,7 @@ sp-session = { version = "2.0.0-rc6", default-features = false }
 sp-std = { version = "2.0.0-rc6", default-features = false }
 sp-transaction-pool = { version = "2.0.0-rc6", default-features = false }
 sp-version = { version = "2.0.0-rc6", default-features = false }
+static_assertions = "1.1.0"
 
 [build-dependencies]
 substrate-wasm-builder-runner = "1.0.6"
